[BugFix] SentMain 뒤로가기시 엉뚱한 값이 불리는 버그 수정
